Move feature helper methods into their own module
