+++ test/test_helper.rb
@@ -7,6 +7,8 @@ require "minitest/rails"
 # to the test group in the Gemfile and uncomment the following:
 require "minitest/rails/capybara"
 
+require_relative "./support/feature_helper"
+
 # Uncomment for awesome colorful output
 # require "minitest/pride"
 
@@ -14,30 +16,5 @@ class ActiveSupport::TestCase
   # Setup all fixtures in test/fixtures/*.yml for all tests in alphabetical order.
   fixtures :all
   # Add more helper methods to be used by all tests here...
-  def create_new_list(title)
-    visit lists_path
-    click_on 'New List'
-    fill_in 'Title', with: title
-    click_on 'Create List'
-  end
-
-  def add_item_to_list(desc)
-    click_on 'New Item'
-    fill_in 'Description', with: desc
-    click_on 'Create List item'
-  end
-
-  def update_list_item_description(desc)
-    within('.list-items') do
-      click_on 'Edit'
-    end
-    fill_in 'Description', with: desc
-    click_on 'Update List item'
-  end
-
-  def update_list_title(title)
-    click_on 'Edit'
-    fill_in 'Title', with: title
-    click_on 'Update List'
-  end
+  include FeatureHelper
 end
